+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 **Enhancements:**
 - **Kubernetes:** Updated scope of the Trident Operator to cluster-scope.
 - Updated GCP driver to allow CVS-Performance volumes as small as 100 GiB.
+- Added `lunsPerFlexvol` option to allow customizing the number of LUNs per FlexVol in the ontap-san-economy driver.
 
 **Beta Features:**